Moved CP utils, cleanup
