+++ addon/components/base/bs-contextual-help.js
@@ -6,8 +6,8 @@ import EmberObject, { computed, observer } from '@ember/object';
 import { cancel, later, next, run, schedule } from '@ember/runloop';
 import transitionEnd from 'ember-bootstrap/utils/transition-end';
 import { getDestinationElement } from '../../utils/dom';
-import usesTransition from 'ember-bootstrap/utils/uses-transition';
-import fastboot from 'ember-bootstrap/utils/fastboot';
+import usesTransition from 'ember-bootstrap/utils/cp/uses-transition';
+import fastboot from 'ember-bootstrap/utils/cp/fastboot';
 
 const InState = EmberObject.extend({
   hover: false,
@@ -22,7 +22,6 @@ function noop() {}
   @class Components.ContextualHelp
   @namespace Components
   @extends Ember.Component
-  @uses Mixins.TransitionSupport
   @private
 */
 let component = Component.extend({
