+++ addon/components/bs3/bs-accordion/item.js
@@ -1,5 +1,5 @@
 import AccordionItem from 'ember-bootstrap/components/base/bs-accordion/item';
-import typeClass from 'ember-bootstrap/utils/type-class';
+import typeClass from 'ember-bootstrap/utils/cp/type-class';
 
 export default AccordionItem.extend({
   classNames: ['panel'],
