+++ addon/components/bs3/bs-progress/bar.js
@@ -1,5 +1,5 @@
 import ProgressBar from 'ember-bootstrap/components/base/bs-progress/bar';
-import typeClass from 'ember-bootstrap/utils/type-class';
+import typeClass from 'ember-bootstrap/utils/cp/type-class';
 
 export default ProgressBar.extend({
   classNameBindings: ['progressBarAnimate:active'],
