+++ addon/components/bs4/bs-form/element/control/checkbox.js
@@ -1,5 +1,5 @@
 import FormElementControlCheckbox from 'ember-bootstrap/components/base/bs-form/element/control/checkbox';
-import formValidationClass from 'ember-bootstrap/utils/form-validation-class';
+import formValidationClass from 'ember-bootstrap/utils/cp/form-validation-class';
 
 export default FormElementControlCheckbox.extend({
   classNames: ['form-check-input'],
