+++ addon/components/bs4/bs-form/element/control/textarea.js
@@ -1,5 +1,5 @@
 import FormElementControlTextarea from 'ember-bootstrap/components/base/bs-form/element/control/textarea';
-import formValidationClass from 'ember-bootstrap/utils/form-validation-class';
+import formValidationClass from 'ember-bootstrap/utils/cp/form-validation-class';
 
 export default FormElementControlTextarea.extend({
   classNameBindings: ['formValidationClass'],
