+++ addon/utils/cp/type-class.js
